rename flavor to forks and chains
