+++ evm/vm/forks/frontier/opcodes.py
