+++ tests/core/block-utils/test_blocks.py
@@ -18,10 +18,10 @@ from evm.db import (
     get_db_backend,
 )
 
-from evm.vm.flavors.frontier.blocks import (
+from evm.vm.forks.frontier.blocks import (
     FrontierBlock,
 )
-from evm.vm.flavors.homestead.blocks import (
+from evm.vm.forks.homestead.blocks import (
     HomesteadBlock,
 )
