+++ tests/json-fixtures/test_blockchain.py
@@ -14,7 +14,7 @@ from evm import (
 from evm.exceptions import (
     ValidationError,
 )
-from evm.vm.flavors import (
+from evm.vm.forks import (
     EIP150VM,
     FrontierVM,
     HomesteadVM,
