+++ tests/json-fixtures/test_transactions.py
@@ -22,7 +22,7 @@ from evm.exceptions import (
 from evm.rlp.headers import (
     BlockHeader,
 )
-from evm.vm.flavors import (
+from evm.vm.forks import (
     MainnetChain,
 )
